ItsFru/backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==3.2.0
boto3==1.35.0
botocore==1.35.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.112.0
h11>=0.12.0,<0.15.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
passlib==1.7.4
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.6
pydantic-settings==2.0.3
httpx>=0.23.0,<0.25.0
pydantic[email]